requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==3.0.3
WTForms==3.1.2
gspread==6.1.2
google-auth==2.29.0
